fixed and reworked scrolling and back buttons
