+++ src/enqueue_places.js
@@ -11,7 +11,7 @@ const MaxCrawledPlacesTracker = require('./max-crawled-places'); // eslint-disab
 const { sleep, log } = Apify.utils;
 const { PLACE_TITLE_SEL, NEXT_BUTTON_SELECTOR, NO_RESULT_XPATH } = require('./consts');
 const { waitForGoogleMapLoader, parseZoomFromUrl } = require('./utils');
-const { parseSearchPlacesResponseBody } = require('./extractors');
+const { parseSearchPlacesResponseBody } = require('./extractors/general');
 const { checkInPolygon } = require('./polygon');
 
 const SEARCH_WAIT_TIME_MS = 30000;
